restructure tests by module add conftest.py to provision fixtures across test modules
